+++ src/symbol/grid_index.ts
@@ -1,3 +1,28 @@
+type QueryArgs = {
+    hitTest: boolean;
+    circle?: {
+        x: number;
+        y: number;
+        radius: number;
+    };
+    seenUids: {
+        box: {
+            [_: number]: boolean;
+        };
+        circle: {
+            [_: number]: boolean;
+        };
+    };
+};
+
+type QueryResult<T> = {
+    key: T;
+    x1: number;
+    y1: number;
+    x2: number;
+    y2: number;
+};
+
 /**
  * GridIndex is a data structure for testing the intersection of
  * circles and rectangles in a 2d plane.
@@ -11,9 +36,9 @@
  *
  * @private
  */
-class GridIndex {
-    circleKeys: Array<any>;
-    boxKeys: Array<any>;
+class GridIndex<T> {
+    circleKeys: Array<T>;
+    boxKeys: Array<T>;
     boxCells: Array<Array<number>>;
     circleCells: Array<Array<number>>;
     bboxes: Array<number>;
@@ -58,7 +83,7 @@ class GridIndex {
         return this.boxKeys.length + this.circleKeys.length;
     }
 
-    insert(key: any, x1: number, y1: number, x2: number, y2: number) {
+    insert(key: T, x1: number, y1: number, x2: number, y2: number) {
         this._forEachCell(x1, y1, x2, y2, this._insertBoxCell, this.boxUid++);
         this.boxKeys.push(key);
         this.bboxes.push(x1);
@@ -67,7 +92,7 @@ class GridIndex {
         this.bboxes.push(y2);
     }
 
-    insertCircle(key: any, x: number, y: number, radius: number) {
+    insertCircle(key: T, x: number, y: number, radius: number) {
         // Insert circle into grid for all cells in the circumscribing square
         // It's more than necessary (by a factor of 4/PI), but fast to insert
         this._forEachCell(x - radius, y - radius, x + radius, y + radius, this._insertCircleCell, this.circleUid++);
@@ -77,22 +102,29 @@ class GridIndex {
         this.circles.push(radius);
     }
 
-    _insertBoxCell(x1: number, y1: number, x2: number, y2: number, cellIndex: number, uid: number) {
+    private _insertBoxCell(x1: number, y1: number, x2: number, y2: number, cellIndex: number, uid: number) {
         this.boxCells[cellIndex].push(uid);
     }
 
-    _insertCircleCell(x1: number, y1: number, x2: number, y2: number, cellIndex: number, uid: number)  {
+    private _insertCircleCell(x1: number, y1: number, x2: number, y2: number, cellIndex: number, uid: number)  {
         this.circleCells[cellIndex].push(uid);
     }
 
-    _query(x1: number, y1: number, x2: number, y2: number, hitTest: boolean, predicate?: any) {
+    private _query(x1: number, y1: number, x2: number, y2: number, hitTest: boolean, predicate?: (key: T) => boolean): Array<QueryResult<T>> {
         if (x2 < 0 || x1 > this.width || y2 < 0 || y1 > this.height) {
-            return hitTest ? false : [];
+            return [];
         }
-        const result = [];
+        const result: Array<QueryResult<T>> = [];
         if (x1 <= 0 && y1 <= 0 && this.width <= x2 && this.height <= y2) {
             if (hitTest) {
-                return true;
+                // Covers the entire grid, so collides with everything
+                return [{
+                    key: null,
+                    x1,
+                    y1,
+                    x2,
+                    y2
+                }];
             }
             for (let boxUid = 0; boxUid < this.boxKeys.length; boxUid++) {
                 result.push({
@@ -115,18 +147,26 @@ class GridIndex {
                     y2: y + radius
                 });
             }
-            return predicate ? result.filter(predicate) : result;
         } else {
-            const queryArgs = {
+            const queryArgs: QueryArgs = {
                 hitTest,
                 seenUids: {box: {}, circle: {}}
             };
             this._forEachCell(x1, y1, x2, y2, this._queryCell, result, queryArgs, predicate);
-            return hitTest ? result.length > 0 : result;
         }
+
+        return result;
     }
 
-    _queryCircle(x: number, y: number, radius: number, hitTest: boolean, predicate?: any) {
+    query(x1: number, y1: number, x2: number, y2: number): Array<QueryResult<T>> {
+        return this._query(x1, y1, x2, y2, false, null);
+    }
+
+    hitTest(x1: number, y1: number, x2: number, y2: number, predicate?: (key: T) => boolean): boolean {
+        return this._query(x1, y1, x2, y2, true, predicate).length > 0;
+    }
+
+    hitTestCircle(x: number, y: number, radius: number, predicate?: (key: T) => boolean): boolean {
         // Insert circle into grid for all cells in the circumscribing square
         // It's more than necessary (by a factor of 4/PI), but fast to insert
         const x1 = x - radius;
@@ -134,59 +174,49 @@ class GridIndex {
         const y1 = y - radius;
         const y2 = y + radius;
         if (x2 < 0 || x1 > this.width || y2 < 0 || y1 > this.height) {
-            return hitTest ? false : [];
+            return false;
         }
 
         // Box query early exits if the bounding box is larger than the grid, but we don't do
         // the equivalent calculation for circle queries because early exit is less likely
         // and the calculation is more expensive
-        const result = [];
-        const queryArgs = {
-            hitTest,
+        const result: boolean[] = [];
+        const queryArgs: QueryArgs = {
+            hitTest: true,
             circle: {x, y, radius},
             seenUids: {box: {}, circle: {}}
         };
         this._forEachCell(x1, y1, x2, y2, this._queryCellCircle, result, queryArgs, predicate);
-        return hitTest ? result.length > 0 : result;
+        return result.length > 0;
     }
 
-    query(x1: number, y1: number, x2: number, y2: number, predicate?: any): Array<any> {
-        return this._query(x1, y1, x2, y2, false, predicate) as any;
-    }
-
-    hitTest(x1: number, y1: number, x2: number, y2: number, predicate?: any): boolean {
-        return this._query(x1, y1, x2, y2, true, predicate) as any;
-    }
-
-    hitTestCircle(x: number, y: number, radius: number, predicate?: any): boolean {
-        return this._queryCircle(x, y, radius, true, predicate) as any;
-    }
-
-    _queryCell(x1: number, y1: number, x2: number, y2: number, cellIndex: number, result: any, queryArgs: any, predicate?: any) {
-        const seenUids = queryArgs.seenUids;
+    private _queryCell(x1: number, y1: number, x2: number, y2: number, cellIndex: number, result: Array<QueryResult<T>>, queryArgs: QueryArgs, predicate?: (key: T) => boolean): boolean {
+        const {seenUids, hitTest} = queryArgs;
         const boxCell = this.boxCells[cellIndex];
+
         if (boxCell !== null) {
             const bboxes = this.bboxes;
             for (const boxUid of boxCell) {
                 if (!seenUids.box[boxUid]) {
                     seenUids.box[boxUid] = true;
                     const offset = boxUid * 4;
+                    const key = this.boxKeys[boxUid];
+
                     if ((x1 <= bboxes[offset + 2]) &&
                         (y1 <= bboxes[offset + 3]) &&
                         (x2 >= bboxes[offset + 0]) &&
                         (y2 >= bboxes[offset + 1]) &&
-                        (!predicate || predicate(this.boxKeys[boxUid]))) {
-                        if (queryArgs.hitTest) {
-                            result.push(true);
+                        (!predicate || predicate(key))) {
+                        result.push({
+                            key,
+                            x1: bboxes[offset],
+                            y1: bboxes[offset + 1],
+                            x2: bboxes[offset + 2],
+                            y2: bboxes[offset + 3]
+                        });
+                        if (hitTest) {
+                            // true return value stops the query after first match
                             return true;
-                        } else {
-                            result.push({
-                                key: this.boxKeys[boxUid],
-                                x1: bboxes[offset],
-                                y1: bboxes[offset + 1],
-                                x2: bboxes[offset + 2],
-                                y2: bboxes[offset + 3]
-                            });
                         }
                     }
                 }
@@ -199,6 +229,8 @@ class GridIndex {
                 if (!seenUids.circle[circleUid]) {
                     seenUids.circle[circleUid] = true;
                     const offset = circleUid * 3;
+                    const key = this.circleKeys[circleUid];
+
                     if (this._circleAndRectCollide(
                         circles[offset],
                         circles[offset + 1],
@@ -207,38 +239,41 @@ class GridIndex {
                         y1,
                         x2,
                         y2) &&
-                        (!predicate || predicate(this.circleKeys[circleUid]))) {
-                        if (queryArgs.hitTest) {
-                            result.push(true);
+                        (!predicate || predicate(key))) {
+                        const x = circles[offset];
+                        const y = circles[offset + 1];
+                        const radius = circles[offset + 2];
+                        result.push({
+                            key,
+                            x1: x - radius,
+                            y1: y - radius,
+                            x2: x + radius,
+                            y2: y + radius
+                        });
+                        if (hitTest) {
+                            // true return value stops the query after first match
                             return true;
-                        } else {
-                            const x = circles[offset];
-                            const y = circles[offset + 1];
-                            const radius = circles[offset + 2];
-                            result.push({
-                                key: this.circleKeys[circleUid],
-                                x1: x - radius,
-                                y1: y - radius,
-                                x2: x + radius,
-                                y2: y + radius
-                            });
                         }
                     }
                 }
             }
         }
+
+        // false return to continue query
+        return false;
     }
 
-    _queryCellCircle(x1: number, y1: number, x2: number, y2: number, cellIndex: number, result: any, queryArgs: any, predicate?: any) {
-        const circle = queryArgs.circle;
-        const seenUids = queryArgs.seenUids;
+    private _queryCellCircle(x1: number, y1: number, x2: number, y2: number, cellIndex: number, result: Array<boolean>, queryArgs: QueryArgs, predicate?: (key:T) => boolean): boolean {
+        const {circle, seenUids} = queryArgs;
         const boxCell = this.boxCells[cellIndex];
+
         if (boxCell !== null) {
             const bboxes = this.bboxes;
             for (const boxUid of boxCell) {
                 if (!seenUids.box[boxUid]) {
                     seenUids.box[boxUid] = true;
                     const offset = boxUid * 4;
+                    const key = this.boxKeys[boxUid];
                     if (this._circleAndRectCollide(
                         circle.x,
                         circle.y,
@@ -247,7 +282,7 @@ class GridIndex {
                         bboxes[offset + 1],
                         bboxes[offset + 2],
                         bboxes[offset + 3]) &&
-                        (!predicate || predicate(this.boxKeys[boxUid]))) {
+                        (!predicate || predicate(key))) {
                         result.push(true);
                         return true;
                     }
@@ -262,6 +297,7 @@ class GridIndex {
                 if (!seenUids.circle[circleUid]) {
                     seenUids.circle[circleUid] = true;
                     const offset = circleUid * 3;
+                    const key = this.circleKeys[circleUid];
                     if (this._circlesCollide(
                         circles[offset],
                         circles[offset + 1],
@@ -269,7 +305,7 @@ class GridIndex {
                         circle.x,
                         circle.y,
                         circle.radius) &&
-                        (!predicate || predicate(this.circleKeys[circleUid]))) {
+                        (!predicate || predicate(key))) {
                         result.push(true);
                         return true;
                     }
@@ -278,7 +314,15 @@ class GridIndex {
         }
     }
 
-    _forEachCell(x1: number, y1: number, x2: number, y2: number, fn: any, arg1: any, arg2?: any, predicate?: any) {
+    private _forEachCell<TArg>(
+        x1: number,
+        y1: number,
+        x2: number,
+        y2: number,
+        fn: (x1: number, y1: number, x2: number, y2: number, cellIndex: number, arg1: TArg, arg2?: QueryArgs, predicate?: (key: T) => boolean) => boolean | void,
+        arg1: TArg,
+        arg2?: QueryArgs,
+        predicate?: (key: T) => boolean) {
         const cx1 = this._convertToXCellCoord(x1);
         const cy1 = this._convertToYCellCoord(y1);
         const cx2 = this._convertToXCellCoord(x2);
@@ -292,22 +336,22 @@ class GridIndex {
         }
     }
 
-    _convertToXCellCoord(x: number) {
+    private _convertToXCellCoord(x: number) {
         return Math.max(0, Math.min(this.xCellCount - 1, Math.floor(x * this.xScale)));
     }
 
-    _convertToYCellCoord(y: number) {
+    private _convertToYCellCoord(y: number) {
         return Math.max(0, Math.min(this.yCellCount - 1, Math.floor(y * this.yScale)));
     }
 
-    _circlesCollide(x1: number, y1: number, r1: number, x2: number, y2: number, r2: number): boolean {
+    private _circlesCollide(x1: number, y1: number, r1: number, x2: number, y2: number, r2: number): boolean {
         const dx = x2 - x1;
         const dy = y2 - y1;
         const bothRadii = r1 + r2;
         return (bothRadii * bothRadii) > (dx * dx + dy * dy);
     }
 
-    _circleAndRectCollide(
+    private _circleAndRectCollide(
       circleX: number,
       circleY: number,
       radius: number,