+++ src/util/test/util.ts
@@ -1,5 +1,6 @@
 import Map from '../../ui/map';
 import {extend} from '../../util/util';
+import Dispatcher from '../../util/dispatcher';
 import gl from 'gl';
 
 export function createMap(options?, callback?) {
@@ -75,4 +76,26 @@ export function setMatchMedia () {
             dispatchEvent: jest.fn(),
         })),
     });
+}
+
+export function getWrapDispatcher () {
+    const wrapDispatcher = (dispatcher) => {
+        return {
+            getActor() {
+                return dispatcher;
+            }
+        } as any as Dispatcher;
+    };
+
+    return wrapDispatcher;
+}
+
+export function getMockDispatcher () {
+    const wrapDispatcher = getWrapDispatcher();
+
+    const mockDispatcher = wrapDispatcher({
+        send () {}
+    });
+
+    return mockDispatcher;
 }